feat(aws): refactor reservationResolver with PriceFilter method
