+++ internal/resources/aws/instance.go
@@ -154,16 +154,18 @@ func (a *Instance) computeCostComponent() *schema.CostComponent {
 		}
 	}
 
+	priceFilter := &schema.PriceFilter{
+		PurchaseOption: strPtr(a.PurchaseOption),
+	}
+
+	var err error
 	if a.ReservedInstanceType != nil {
 		resolver := newEc2ReservationResolver(strVal(a.ReservedInstanceTerm), strVal(a.ReservedInstancePaymentOption), strVal(a.ReservedInstanceType))
-		valid, err := resolver.Validate()
-		if err != "" {
-			log.Warnf(err)
+		priceFilter, err = resolver.PriceFilter()
+		if err != nil {
+			log.Warnf(err.Error())
 		}
-		if valid {
-			purchaseOptionLabel = "reserved"
-			return a.reservedInstanceCostComponent(osLabel, osFilterVal, purchaseOptionLabel, resolver)
-		}
+		purchaseOptionLabel = "reserved"
 	}
 
 	return &schema.CostComponent{
@@ -185,38 +187,7 @@ func (a *Instance) computeCostComponent() *schema.CostComponent {
 				{Key: "capacitystatus", Value: strPtr("Used")},
 			},
 		},
-		PriceFilter: &schema.PriceFilter{
-			PurchaseOption: strPtr(a.PurchaseOption),
-		},
-	}
-}
-
-func (a *Instance) reservedInstanceCostComponent(osLabel, osFilterVal, purchaseOptionLabel string, resolver reservationResolver) *schema.CostComponent {
-
-	return &schema.CostComponent{
-		Name:           fmt.Sprintf("Instance usage (%s, %s, %s)", osLabel, purchaseOptionLabel, a.InstanceType),
-		Unit:           "hours",
-		UnitMultiplier: decimal.NewFromInt(1),
-		HourlyQuantity: decimalPtr(decimal.NewFromInt(1)),
-		ProductFilter: &schema.ProductFilter{
-			VendorName:    strPtr("aws"),
-			Region:        strPtr(a.Region),
-			Service:       strPtr("AmazonEC2"),
-			ProductFamily: strPtr("Compute Instance"),
-			AttributeFilters: []*schema.AttributeFilter{
-				{Key: "instanceType", Value: strPtr(a.InstanceType)},
-				{Key: "tenancy", Value: strPtr(a.Tenancy)},
-				{Key: "operatingSystem", Value: strPtr(osFilterVal)},
-				{Key: "preInstalledSw", Value: strPtr("NA")},
-				{Key: "capacitystatus", Value: strPtr("Used")},
-			},
-		},
-		PriceFilter: &schema.PriceFilter{
-			StartUsageAmount:   strPtr("0"),
-			TermOfferingClass:  a.ReservedInstanceType,
-			TermLength:         strPtr(resolver.Term()),
-			TermPurchaseOption: strPtr(resolver.PaymentOption()),
-		},
+		PriceFilter: priceFilter,
 	}
 }
 
@@ -299,4 +270,42 @@ func (a *Instance) cpuCreditCostComponent(instanceFamily string) *schema.CostCom
 			},
 		},
 	}
+}
+
+// EC2 implementation of reservationResolver
+type ec2ReservationResolver struct {
+	term              string
+	paymentOption     string
+	termOfferingClass string
+}
+
+func newEc2ReservationResolver(term, paymentOption, termOfferingClass string) reservationResolver {
+	return &ec2ReservationResolver{
+		term:              term,
+		paymentOption:     paymentOption,
+		termOfferingClass: termOfferingClass,
+	}
+}
+
+func (r ec2ReservationResolver) PriceFilter() (*schema.PriceFilter, error) {
+	termLength := reservedTermsMapping[r.term]
+	purchaseOption := reservedPaymentOptionMapping[r.paymentOption]
+	validTypes := []string{"convertible", "standard"}
+	if !stringInSlice(validTypes, r.termOfferingClass) {
+		return nil, fmt.Errorf("Invalid reserved_instance_type, ignoring reserved options. Expected: convertible, standard. Got: %s", r.termOfferingClass)
+	}
+	validTerms := sliceOfKeysFromMap(reservedTermsMapping)
+	if !stringInSlice(validTerms, r.term) {
+		return nil, fmt.Errorf("Invalid reserved_instance_term, ignoring reserved options. Expected: %s. Got: %s", strings.Join(validTerms, ", "), r.term)
+	}
+	validOptions := sliceOfKeysFromMap(reservedPaymentOptionMapping)
+	if !stringInSlice(validOptions, r.paymentOption) {
+		return nil, fmt.Errorf("Invalid reserved_instance_payment_option, ignoring reserved options. Expected: %s. Got: %s", strings.Join(validOptions, ", "), r.paymentOption)
+	}
+	return &schema.PriceFilter{
+		StartUsageAmount:   strPtr("0"),
+		TermOfferingClass:  strPtr(r.termOfferingClass),
+		TermLength:         strPtr(termLength),
+		TermPurchaseOption: strPtr(purchaseOption),
+	}, nil
 }
